move mutex test workflow to the tests dir
