--- app/agent/activities/mutex/test-activity.ts
@@ -1,6 +1,0 @@
-import type { CreateActivity } from "../types";
-
-export type MutexTestActivity = () => Promise<number>;
-export const mutexTest: CreateActivity<MutexTestActivity> = () => async () => {
-  return 0;
-};
