+++ app/agent/workflows/index.ts
@@ -8,7 +8,6 @@ import {
 } from "@temporalio/workflow";
 
 import { lockWorkflow } from "./mutex";
-import { testLock } from "./mutex/test-workflow";
 import { runBuildfsAndPrebuilds, runPrebuild, runBuildfs, scheduleNewPrebuild } from "./prebuild";
 import { runWaitForWorkflow } from "./wait-for-workflow";
 import {
@@ -34,7 +33,6 @@ export {
   monitorWorkspaceInstance,
   runDeleteWorkspace,
   lockWorkflow,
-  testLock,
   scheduleNewPrebuild,
   runWaitForWorkflow,
 };
