+++ app/agent/workflows/tests/activities.ts
@@ -1,5 +1,6 @@
 import type { Prisma } from "@prisma/client";
 
+import { mutexTest } from "./mutex/activity";
 import { cancellationTest } from "./wait-for-workflow/activity";
 
 import type { Activities } from "~/agent/activities/list";
@@ -8,6 +9,7 @@ import type { AgentInjector } from "~/agent/agent-injector";
 
 const testActivities = {
   cancellationTest,
+  mutexTest,
 } as const;
 
 export type TestActivities = Activities & typeof testActivities;
