+++ app/agent/workflows/tests/mutex/activity.ts
@@ -1,0 +1,3 @@
+export const mutexTest = async (): Promise<number> => {
+  return 0;
+};
