+++ app/agent/workflows/tests/workflows.test.ts
@@ -3,10 +3,10 @@ import { Mutex } from "async-mutex";
 import { v4 as uuidv4 } from "uuid";
 
 import { prepareTests } from "./utils";
-import { cancellationTestWorkflow } from "./wait-for-workflow/workflow";
+import { testLock, cancellationTestWorkflow } from "./workflows";
 
 import { withActivityHeartbeat } from "~/agent/activities/utils";
-import { unwrap } from "~/utils.shared";
+import { unwrap, waitForPromises } from "~/utils.shared";
 
 const { provideTestActivities } = prepareTests();
 
@@ -73,6 +73,63 @@ test.concurrent(
       } finally {
         for (const { release } of locks.values()) {
           release();
+        }
+      }
+    });
+  }),
+);
+
+test.concurrent(
+  "withLock",
+  provideTestActivities(async ({ createWorker }) => {
+    let counter = 0;
+    const { worker, client, taskQueue } = await createWorker({
+      activityOverrides: {
+        mutexTest: async () => {
+          return counter++;
+        },
+      },
+    });
+
+    await worker.runUntil(async () => {
+      const results = await waitForPromises(
+        [0, 1, 2].map(() =>
+          client.workflow.execute(testLock, {
+            workflowId: uuidv4(),
+            taskQueue,
+            retry: { maximumAttempts: 1 },
+            args: [],
+          }),
+        ),
+      );
+      expect(results).toHaveLength(3);
+      const parseResult = (workflowResult: string[], i: number): number => {
+        const acquire = workflowResult[i];
+        const activityResult = workflowResult[i + 1];
+        const release = workflowResult[i + 2];
+        expect(acquire).toMatch(/^acquire-\d+$/);
+        expect(activityResult).toMatch(/^result-\d+-\d+$/);
+        expect(release).toMatch(/^release-\d+$/);
+        const acquireId = parseInt(acquire.split("-")[1]);
+        const splitResult = activityResult.split("-");
+        const activityResultId = parseInt(splitResult[1]);
+        const activityResultValue = parseInt(splitResult[2]);
+        const releaseId = parseInt(release.split("-")[1]);
+        // Here we check that the inner lock works
+        expect(acquireId).toBe(releaseId);
+        expect(activityResultId).toBe(releaseId);
+        return activityResultValue;
+      };
+      for (const result of results) {
+        expect(result).toHaveLength(15);
+        const activityResults = Array.from({ length: result.length / 3 }).map((_, i) =>
+          parseResult(result, i * 3),
+        );
+        // Here we check that the outer lock works
+        for (let i = 0; i < activityResults.length - 1; i++) {
+          const curActivityResult = activityResults[i];
+          const nextActivityResult = activityResults[i + 1];
+          expect(curActivityResult + 1).toBe(nextActivityResult);
         }
       }
     });
