+++ app/agent/workflows/tests/workflows.ts
@@ -1,3 +1,4 @@
 export * from "~/agent/workflows";
 
 export { cancellationTestWorkflow } from "./wait-for-workflow/workflow";
+export { testLock } from "./mutex/workflow";
